refactor: Remove and refactor code
refactor: Create a prototype time graph element

Move some state content over

refactor: rename folder

refactor: More renaming and pruning of old unused code.

Create a prototype time graph element

scrollable table

Temporarily add dead code allows

Minor refactorings

refactor: Move current widgets to the new element system

squash this - WIP of re-refactor

Revert "squash this - WIP of re-refactor"

This reverts commit a3ce25732663863b1ba2ef8a764c57572406014f.

Revert "refactor: Move current widgets to the new element system"

This reverts commit c7218472c2ce6d1bebf5550aaf53c171d3cd7c3e.

WIP re-refactor

Revert "refactor: Create a prototype time graph element"

This reverts commit bd95ca8b73623c9df1dab03e0c51b31cb9b99c4d.
