+++ src/app.rs
@@ -1,8 +1,14 @@
+pub mod data_farmer;
+pub mod data_harvester;
+pub mod filter;
+pub mod layout_manager;
+mod process_killer;
+pub mod query;
+pub mod widget_states;
+
 use std::{
     cmp::{max, min},
     collections::HashMap,
-    // io::Write,
-    path::PathBuf,
     time::Instant,
 };
 
@@ -13,25 +19,16 @@ use typed_builder::*;
 
 use data_farmer::*;
 use data_harvester::{processes, temperature};
+pub use filter::*;
 use layout_manager::*;
-pub use states::*;
+pub use widget_states::*;
 
 use crate::{
     canvas, constants,
-    options::Config,
-    options::ConfigFlags,
-    options::WidgetIdEnabled,
     units::data_units::DataUnit,
     utils::error::{BottomError, Result},
     Pid,
 };
-
-pub mod data_farmer;
-pub mod data_harvester;
-pub mod layout_manager;
-mod process_killer;
-pub mod query;
-pub mod states;
 
 const MAX_SEARCH_LENGTH: usize = 200;
 
@@ -68,23 +65,8 @@ pub struct AppConfigFields {
     pub network_use_binary_prefix: bool,
 }
 
-/// For filtering out information
-#[derive(Debug, Clone)]
-pub struct DataFilters {
-    pub disk_filter: Option<Filter>,
-    pub mount_filter: Option<Filter>,
-    pub temp_filter: Option<Filter>,
-    pub net_filter: Option<Filter>,
-}
-
-#[derive(Debug, Clone)]
-pub struct Filter {
-    pub is_list_ignored: bool,
-    pub list: Vec<regex::Regex>,
-}
-
 #[derive(TypedBuilder)]
-pub struct App {
+pub struct AppState {
     #[builder(default = false, setter(skip))]
     awaiting_second_char: bool,
 
@@ -127,12 +109,6 @@ pub struct App {
     #[builder(default = false, setter(skip))]
     pub basic_mode_use_percent: bool,
 
-    #[builder(default = false, setter(skip))]
-    pub is_config_open: bool,
-
-    #[builder(default = false, setter(skip))]
-    pub did_config_fail_to_save: bool,
-
     #[cfg(target_family = "unix")]
     #[builder(default, setter(skip))]
     pub user_table: processes::UserTable,
@@ -150,8 +126,6 @@ pub struct App {
     pub current_widget: BottomWidget,
     pub used_widgets: UsedWidgets,
     pub filters: DataFilters,
-    pub config: Config,
-    pub config_path: Option<PathBuf>,
 }
 
 #[cfg(target_os = "windows")]
@@ -161,7 +135,7 @@ const MAX_SIGNAL: usize = 64;
 #[cfg(target_os = "macos")]
 const MAX_SIGNAL: usize = 31;
 
-impl App {
+impl AppState {
     pub fn reset(&mut self) {
         // Reset multi
         self.reset_multi_tap_keys();
@@ -218,8 +192,6 @@ impl App {
             }
 
             self.is_force_redraw = true;
-        } else if self.is_config_open {
-            self.close_config_screen();
         } else {
             match self.current_widget.widget_type {
                 BottomWidgetType::Proc => {
@@ -297,7 +269,7 @@ impl App {
     }
 
     fn ignore_normal_keybinds(&self) -> bool {
-        self.is_config_open || self.is_in_dialog()
+        self.is_in_dialog()
     }
 
     pub fn on_tab(&mut self) {
@@ -507,7 +479,6 @@ impl App {
 
     pub fn toggle_ignore_case(&mut self) {
         let is_in_search_widget = self.is_in_search_widget();
-        let mut is_case_sensitive: Option<bool> = None;
         if let Some(proc_widget_state) = self
             .proc_state
             .widget_states
@@ -519,50 +490,12 @@ impl App {
                     .search_toggle_ignore_case();
                 proc_widget_state.update_query();
                 self.proc_state.force_update = Some(self.current_widget.widget_id - 1);
-
-                // Remember, it's the opposite (ignoring case is case "in"sensitive)
-                is_case_sensitive = Some(!proc_widget_state.process_search_state.is_ignoring_case);
             }
-        }
-
-        // Also toggle it in the config file if we actually changed it.
-        if let Some(is_ignoring_case) = is_case_sensitive {
-            if let Some(flags) = &mut self.config.flags {
-                if let Some(map) = &mut flags.search_case_enabled_widgets_map {
-                    // Just update the map.
-                    let mapping = map.entry(self.current_widget.widget_id - 1).or_default();
-                    *mapping = is_ignoring_case;
-
-                    flags.search_case_enabled_widgets =
-                        Some(WidgetIdEnabled::create_from_hashmap(&map));
-                } else {
-                    // Map doesn't exist yet... initialize ourselves.
-                    let mut map = HashMap::default();
-                    map.insert(self.current_widget.widget_id - 1, is_ignoring_case);
-                    flags.search_case_enabled_widgets =
-                        Some(WidgetIdEnabled::create_from_hashmap(&map));
-                    flags.search_case_enabled_widgets_map = Some(map);
-                }
-            } else {
-                // Must initialize it ourselves...
-                let mut map = HashMap::default();
-                map.insert(self.current_widget.widget_id - 1, is_ignoring_case);
-
-                self.config.flags = Some(
-                    ConfigFlags::builder()
-                        .search_case_enabled_widgets(WidgetIdEnabled::create_from_hashmap(&map))
-                        .search_case_enabled_widgets_map(map)
-                        .build(),
-                );
-            }
-
-            // self.did_config_fail_to_save = self.update_config_file().is_err();
         }
     }
 
     pub fn toggle_search_whole_word(&mut self) {
         let is_in_search_widget = self.is_in_search_widget();
-        let mut is_searching_whole_word: Option<bool> = None;
         if let Some(proc_widget_state) = self
             .proc_state
             .widget_states
@@ -574,55 +507,12 @@ impl App {
                     .search_toggle_whole_word();
                 proc_widget_state.update_query();
                 self.proc_state.force_update = Some(self.current_widget.widget_id - 1);
-
-                is_searching_whole_word = Some(
-                    proc_widget_state
-                        .process_search_state
-                        .is_searching_whole_word,
-                );
             }
-        }
-
-        // Also toggle it in the config file if we actually changed it.
-        if let Some(is_searching_whole_word) = is_searching_whole_word {
-            if let Some(flags) = &mut self.config.flags {
-                if let Some(map) = &mut flags.search_whole_word_enabled_widgets_map {
-                    // Just update the map.
-                    let mapping = map.entry(self.current_widget.widget_id - 1).or_default();
-                    *mapping = is_searching_whole_word;
-
-                    flags.search_whole_word_enabled_widgets =
-                        Some(WidgetIdEnabled::create_from_hashmap(&map));
-                } else {
-                    // Map doesn't exist yet... initialize ourselves.
-                    let mut map = HashMap::default();
-                    map.insert(self.current_widget.widget_id - 1, is_searching_whole_word);
-                    flags.search_whole_word_enabled_widgets =
-                        Some(WidgetIdEnabled::create_from_hashmap(&map));
-                    flags.search_whole_word_enabled_widgets_map = Some(map);
-                }
-            } else {
-                // Must initialize it ourselves...
-                let mut map = HashMap::default();
-                map.insert(self.current_widget.widget_id - 1, is_searching_whole_word);
-
-                self.config.flags = Some(
-                    ConfigFlags::builder()
-                        .search_whole_word_enabled_widgets(WidgetIdEnabled::create_from_hashmap(
-                            &map,
-                        ))
-                        .search_whole_word_enabled_widgets_map(map)
-                        .build(),
-                );
-            }
-
-            // self.did_config_fail_to_save = self.update_config_file().is_err();
         }
     }
 
     pub fn toggle_search_regex(&mut self) {
         let is_in_search_widget = self.is_in_search_widget();
-        let mut is_searching_with_regex: Option<bool> = None;
         if let Some(proc_widget_state) = self
             .proc_state
             .widget_states
@@ -632,47 +522,7 @@ impl App {
                 proc_widget_state.process_search_state.search_toggle_regex();
                 proc_widget_state.update_query();
                 self.proc_state.force_update = Some(self.current_widget.widget_id - 1);
-
-                is_searching_with_regex = Some(
-                    proc_widget_state
-                        .process_search_state
-                        .is_searching_with_regex,
-                );
             }
-        }
-
-        // Also toggle it in the config file if we actually changed it.
-        if let Some(is_searching_whole_word) = is_searching_with_regex {
-            if let Some(flags) = &mut self.config.flags {
-                if let Some(map) = &mut flags.search_regex_enabled_widgets_map {
-                    // Just update the map.
-                    let mapping = map.entry(self.current_widget.widget_id - 1).or_default();
-                    *mapping = is_searching_whole_word;
-
-                    flags.search_regex_enabled_widgets =
-                        Some(WidgetIdEnabled::create_from_hashmap(&map));
-                } else {
-                    // Map doesn't exist yet... initialize ourselves.
-                    let mut map = HashMap::default();
-                    map.insert(self.current_widget.widget_id - 1, is_searching_whole_word);
-                    flags.search_regex_enabled_widgets =
-                        Some(WidgetIdEnabled::create_from_hashmap(&map));
-                    flags.search_regex_enabled_widgets_map = Some(map);
-                }
-            } else {
-                // Must initialize it ourselves...
-                let mut map = HashMap::default();
-                map.insert(self.current_widget.widget_id - 1, is_searching_whole_word);
-
-                self.config.flags = Some(
-                    ConfigFlags::builder()
-                        .search_regex_enabled_widgets(WidgetIdEnabled::create_from_hashmap(&map))
-                        .search_regex_enabled_widgets_map(map)
-                        .build(),
-                );
-            }
-
-            // self.did_config_fail_to_save = self.update_config_file().is_err();
         }
     }
 
@@ -910,8 +760,7 @@ impl App {
     }
 
     pub fn on_up_key(&mut self) {
-        if self.is_config_open {
-        } else if !self.is_in_dialog() {
+        if !self.is_in_dialog() {
             self.decrement_position_count();
         } else if self.help_dialog_state.is_showing_help {
             self.help_scroll_up();
@@ -932,8 +781,7 @@ impl App {
     }
 
     pub fn on_down_key(&mut self) {
-        if self.is_config_open {
-        } else if !self.is_in_dialog() {
+        if !self.is_in_dialog() {
             self.increment_position_count();
         } else if self.help_dialog_state.is_showing_help {
             self.help_scroll_down();
@@ -954,8 +802,7 @@ impl App {
     }
 
     pub fn on_left_key(&mut self) {
-        if self.is_config_open {
-        } else if !self.is_in_dialog() {
+        if !self.is_in_dialog() {
             match self.current_widget.widget_type {
                 BottomWidgetType::ProcSearch => {
                     let is_in_search_widget = self.is_in_search_widget();
@@ -1026,8 +873,7 @@ impl App {
     }
 
     pub fn on_right_key(&mut self) {
-        if self.is_config_open {
-        } else if !self.is_in_dialog() {
+        if !self.is_in_dialog() {
             match self.current_widget.widget_type {
                 BottomWidgetType::ProcSearch => {
                     let is_in_search_widget = self.is_in_search_widget();
@@ -1163,7 +1009,6 @@ impl App {
                     }
                 }
             }
-        } else if self.is_config_open {
         }
     }
 
@@ -1210,7 +1055,6 @@ impl App {
                     }
                 }
             }
-        } else if self.is_config_open {
         }
     }
 
@@ -1464,7 +1308,6 @@ impl App {
                 'G' => self.skip_to_last(),
                 _ => {}
             }
-        } else if self.is_config_open {
         }
     }
 
@@ -1646,39 +1489,6 @@ impl App {
     }
 
     pub fn on_space(&mut self) {}
-
-    pub fn open_config_screen(&mut self) {
-        self.is_config_open = true;
-        self.is_force_redraw = true;
-    }
-
-    pub fn close_config_screen(&mut self) {
-        self.is_config_open = false;
-        self.is_force_redraw = true;
-    }
-
-    /// TODO: Disabled.
-    /// Call this whenever the config value is updated!
-    // fn update_config_file(&mut self) -> anyhow::Result<()> {
-    //     if self.app_config_fields.no_write {
-    //         // debug!("No write enabled.  Config will not be written.");
-    //         // Don't write!
-    //         // FIXME: [CONFIG] This should be made VERY clear to the user... make a thing saying "it will not write due to no_write option"
-    //         Ok(())
-    //     } else if let Some(config_path) = &self.config_path {
-    //         // Update
-    //         // debug!("Updating config file - writing to: {:?}", config_path);
-    //         std::fs::File::create(config_path)?
-    //             .write_all(self.config.get_config_as_bytes()?.as_ref())?;
-    //         Ok(())
-    //     } else {
-    //         // FIXME: [CONFIG] Put an actual error message?
-    //         Err(anyhow::anyhow!(
-    //             "Config path was missing, please try restarting bottom..."
-    //         ))
-    //     }
-    //     Ok(())
-    // }
 
     pub fn kill_highlighted_process(&mut self) -> Result<()> {
         if let BottomWidgetType::Proc = self.current_widget.widget_type {
@@ -2238,7 +2048,6 @@ impl App {
                 _ => {}
             }
             self.reset_multi_tap_keys();
-        } else if self.is_config_open {
         } else if self.help_dialog_state.is_showing_help {
             self.help_dialog_state.scroll_state.current_scroll_index = 0;
         } else if self.delete_dialog_state.is_showing_dd {
@@ -2317,7 +2126,6 @@ impl App {
                 _ => {}
             }
             self.reset_multi_tap_keys();
-        } else if self.is_config_open {
         } else if self.help_dialog_state.is_showing_help {
             self.help_dialog_state.scroll_state.current_scroll_index = self
                 .help_dialog_state
