+++ src/app/widget_states/battery_state.rs
@@ -1,0 +1,25 @@
+use std::collections::HashMap;
+
+#[derive(Default)]
+pub struct BatteryWidgetState {
+    pub currently_selected_battery_index: usize,
+    pub tab_click_locs: Option<Vec<((u16, u16), (u16, u16))>>,
+}
+
+pub struct BatteryState {
+    pub widget_states: HashMap<u64, BatteryWidgetState>,
+}
+
+impl BatteryState {
+    pub fn init(widget_states: HashMap<u64, BatteryWidgetState>) -> Self {
+        BatteryState { widget_states }
+    }
+
+    pub fn get_mut_widget_state(&mut self, widget_id: u64) -> Option<&mut BatteryWidgetState> {
+        self.widget_states.get_mut(&widget_id)
+    }
+
+    pub fn get_widget_state(&self, widget_id: u64) -> Option<&BatteryWidgetState> {
+        self.widget_states.get(&widget_id)
+    }
+}
