+++ src/app/widget_states/cpu_state.rs
@@ -1,0 +1,47 @@
+use std::{collections::HashMap, time::Instant};
+
+use super::{AppScrollWidgetState, CanvasTableWidthState};
+
+pub struct CpuWidgetState {
+    pub current_display_time: u64,
+    pub is_legend_hidden: bool,
+    pub autohide_timer: Option<Instant>,
+    pub scroll_state: AppScrollWidgetState,
+    pub is_multi_graph_mode: bool,
+    pub table_width_state: CanvasTableWidthState,
+}
+
+impl CpuWidgetState {
+    pub fn init(current_display_time: u64, autohide_timer: Option<Instant>) -> Self {
+        CpuWidgetState {
+            current_display_time,
+            is_legend_hidden: false,
+            autohide_timer,
+            scroll_state: AppScrollWidgetState::default(),
+            is_multi_graph_mode: false,
+            table_width_state: CanvasTableWidthState::default(),
+        }
+    }
+}
+
+pub struct CpuState {
+    pub force_update: Option<u64>,
+    pub widget_states: HashMap<u64, CpuWidgetState>,
+}
+
+impl CpuState {
+    pub fn init(widget_states: HashMap<u64, CpuWidgetState>) -> Self {
+        CpuState {
+            force_update: None,
+            widget_states,
+        }
+    }
+
+    pub fn get_mut_widget_state(&mut self, widget_id: u64) -> Option<&mut CpuWidgetState> {
+        self.widget_states.get_mut(&widget_id)
+    }
+
+    pub fn get_widget_state(&self, widget_id: u64) -> Option<&CpuWidgetState> {
+        self.widget_states.get(&widget_id)
+    }
+}
