+++ src/app/widget_states/disk_state.rs
@@ -1,0 +1,35 @@
+use std::collections::HashMap;
+
+use super::{AppScrollWidgetState, CanvasTableWidthState};
+
+pub struct DiskWidgetState {
+    pub scroll_state: AppScrollWidgetState,
+    pub table_width_state: CanvasTableWidthState,
+}
+
+impl DiskWidgetState {
+    pub fn init() -> Self {
+        DiskWidgetState {
+            scroll_state: AppScrollWidgetState::default(),
+            table_width_state: CanvasTableWidthState::default(),
+        }
+    }
+}
+
+pub struct DiskState {
+    pub widget_states: HashMap<u64, DiskWidgetState>,
+}
+
+impl DiskState {
+    pub fn init(widget_states: HashMap<u64, DiskWidgetState>) -> Self {
+        DiskState { widget_states }
+    }
+
+    pub fn get_mut_widget_state(&mut self, widget_id: u64) -> Option<&mut DiskWidgetState> {
+        self.widget_states.get_mut(&widget_id)
+    }
+
+    pub fn get_widget_state(&self, widget_id: u64) -> Option<&DiskWidgetState> {
+        self.widget_states.get(&widget_id)
+    }
+}
