+++ src/app/widget_states/graph_state.rs
@@ -1,0 +1,1 @@
+//! States for a graph widget.
