+++ src/app/widget_states/mem_state.rs
@@ -1,0 +1,37 @@
+use std::{collections::HashMap, time::Instant};
+
+pub struct MemWidgetState {
+    pub current_display_time: u64,
+    pub autohide_timer: Option<Instant>,
+}
+
+impl MemWidgetState {
+    pub fn init(current_display_time: u64, autohide_timer: Option<Instant>) -> Self {
+        MemWidgetState {
+            current_display_time,
+            autohide_timer,
+        }
+    }
+}
+
+pub struct MemState {
+    pub force_update: Option<u64>,
+    pub widget_states: HashMap<u64, MemWidgetState>,
+}
+
+impl MemState {
+    pub fn init(widget_states: HashMap<u64, MemWidgetState>) -> Self {
+        MemState {
+            force_update: None,
+            widget_states,
+        }
+    }
+
+    pub fn get_mut_widget_state(&mut self, widget_id: u64) -> Option<&mut MemWidgetState> {
+        self.widget_states.get_mut(&widget_id)
+    }
+
+    pub fn get_widget_state(&self, widget_id: u64) -> Option<&MemWidgetState> {
+        self.widget_states.get(&widget_id)
+    }
+}
