+++ src/app/widget_states/table_state.rs
@@ -1,0 +1,1 @@
+//! States for a table widget.
