+++ src/app/widget_states/temp_state.rs
@@ -1,0 +1,35 @@
+use std::collections::HashMap;
+
+use super::{AppScrollWidgetState, CanvasTableWidthState};
+
+pub struct TempWidgetState {
+    pub scroll_state: AppScrollWidgetState,
+    pub table_width_state: CanvasTableWidthState,
+}
+
+impl TempWidgetState {
+    pub fn init() -> Self {
+        TempWidgetState {
+            scroll_state: AppScrollWidgetState::default(),
+            table_width_state: CanvasTableWidthState::default(),
+        }
+    }
+}
+
+pub struct TempState {
+    pub widget_states: HashMap<u64, TempWidgetState>,
+}
+
+impl TempState {
+    pub fn init(widget_states: HashMap<u64, TempWidgetState>) -> Self {
+        TempState { widget_states }
+    }
+
+    pub fn get_mut_widget_state(&mut self, widget_id: u64) -> Option<&mut TempWidgetState> {
+        self.widget_states.get_mut(&widget_id)
+    }
+
+    pub fn get_widget_state(&self, widget_id: u64) -> Option<&TempWidgetState> {
+        self.widget_states.get(&widget_id)
+    }
+}
