--- src/canvas/screens.rs
@@ -1,3 +1,0 @@
-pub mod config_screen;
-
-pub use config_screen::*;
