+++ src/canvas/widgets/basic_table_arrows.rs
@@ -1,5 +1,5 @@
 use crate::{
-    app::{layout_manager::BottomWidgetType, App},
+    app::{layout_manager::BottomWidgetType, AppState},
     canvas::Painter,
 };
 
@@ -14,13 +14,13 @@ use tui::{
 
 pub trait BasicTableArrows {
     fn draw_basic_table_arrows<B: Backend>(
-        &self, f: &mut Frame<'_, B>, app_state: &mut App, draw_loc: Rect, widget_id: u64,
+        &self, f: &mut Frame<'_, B>, app_state: &mut AppState, draw_loc: Rect, widget_id: u64,
     );
 }
 
 impl BasicTableArrows for Painter {
     fn draw_basic_table_arrows<B: Backend>(
-        &self, f: &mut Frame<'_, B>, app_state: &mut App, draw_loc: Rect, widget_id: u64,
+        &self, f: &mut Frame<'_, B>, app_state: &mut AppState, draw_loc: Rect, widget_id: u64,
     ) {
         if let Some(current_table) = app_state.widget_map.get(&widget_id) {
             let current_table = if let BottomWidgetType::ProcSort = current_table.widget_type {
