+++ src/canvas/widgets/disk_table.rs
@@ -29,15 +29,15 @@ static DISK_HEADERS_LENS: Lazy<Vec<u16>> = Lazy::new(|| {
 
 pub trait DiskTableWidget {
     fn draw_disk_table<B: Backend>(
-        &self, f: &mut Frame<'_, B>, app_state: &mut app::App, draw_loc: Rect, draw_border: bool,
-        widget_id: u64,
+        &self, f: &mut Frame<'_, B>, app_state: &mut app::AppState, draw_loc: Rect,
+        draw_border: bool, widget_id: u64,
     );
 }
 
 impl DiskTableWidget for Painter {
     fn draw_disk_table<B: Backend>(
-        &self, f: &mut Frame<'_, B>, app_state: &mut app::App, draw_loc: Rect, draw_border: bool,
-        widget_id: u64,
+        &self, f: &mut Frame<'_, B>, app_state: &mut app::AppState, draw_loc: Rect,
+        draw_border: bool, widget_id: u64,
     ) {
         let recalculate_column_widths = app_state.should_get_widget_bounds();
         if let Some(disk_widget_state) = app_state.disk_state.widget_states.get_mut(&widget_id) {
